Add AddApplicantParser class to ReCLIne
AddApplicantParser is used to detect and parse an AddApplicant input,
and return an AddApplicant Command.

Let's add an AddApplicantParser class that parses the input for the AddApplicant command.
Let's add a dummy AddApplicant command class and Applicant class so that we can test our
parser.
Let's also add the required changes to the methods in AddressBookParser, ParserUtil,
and AddressBook.

This changes make it easier to integrate when we merge the other classes and methods, and
allows us to test the functionality of the Parser.

We assume here that the required fields when adding an Applicant in this commit is
all fields except Id and Status.
Further changes to be made to the AddApplicantParser when other attribute
classes are merged to the repo.