Add AddApplicant class to ReCLIne
AddApplicant is used to add an Applicant to an ApplicantList in the AddressBook object.

Let's add an AddApplicant class with and execute method that adds an Applicant to the list.
Let's add an Applicant class with required attributes, using String for classes that have
not been merged.
Let's modify ModelManager and Model interface to include an ArrayList as the stand in for the
ApplicantList, and also the methods to add and check for duplicate Applicant.

This changes make it easier to integrate when we merge the other classes and methods, and
allows us to test the functionality of the AddApplicant class up to the execute method affecting
the ModelManager.

We assume here that the required fields when adding an Applicant in this commit is
all fields except Id and Status.
Further changes to be made to the AddApplicant, Applicant and ModelManager when other attribute
classes are merged to the repo.