Add storage classes to store ApplicantList
We need to be able to store the Applicant List that contains the new Applicant class we have created.

Let's add a JsonAdaptedApplicant so that we can convert an Applicant class to its JSON form.
Lets also modify the toModelType method in JsonSerializableAddressBook to be able to convert JSON with ApplicantList.
Lets also modify the tests in JsonSerializableAddressBookTest to be able to take in a JSON list with Applicants.
Lets also add validApplicantsAddressBook.json, invalidApplicantsAddressBook.json, and duplicateApplicantsAddressBook.json files for testing.

This changes allow us to save an Applicant List in JSON format alongside the Person list

We assume here that we are keeping the Person list.
Further changes to be made to the storage methods when other attributes are added/removed (Jobid, Applicationstatus, Id).