Implement AddJob feature in ReCLIne
We need to be able to add new Jobs to ReCLIne.

Lets add an AddJob command class to add a Job to the UniqueJobList in ReCLIne

Lets also add an AddJobParser to parser user input when adding a new Job.

Lets make AddressBook store a IdCount attribute which is used to generate unique Ids for new Jobs.

Lets create a method in ModelManager which gets the current idcount, returns it to the AddJob command to create a Job with that IdCount, then increment the IdCount in Addressbook.

Lastly lets change storage classes to store and read the new IdCount attribute.

This changes allow us to add a new Job to ReCLIne, and have a unique IdCount that is stored in the save data file.